clients/requirements.txt
opencv-python
requests
python-socketio[client]==5.11.1